Added JavaDoc. Minor visibility change.
Mostly just adding comments. It was helpful to look at the code again
and reflect on it as a learning exercise. I pretty proud of this; it
isn't particularly impressive, but it was my first shot at working with
JavaFX and I still understand what it is doing, which I think means I
did a pretty good job.