Migrate the build system over to Gradle
This is much less of "migration" as an introduction. There
was no build system before, and I was relying on the IDE to
do the build for me. Which wasn't good, especially as I actually
do have a compile dependency (and now test dependencies). I went
with Gradle because I haven't used it before and I figured I'd
try it out.