Begin introducing MVC pattern
Beginning to enforce separation of concerns a little bit here.
In particular, I'm pretty happy with how the Model looks at this
moment, so I think that is likely to be finished. The outline of
the view is done, sans any of the tables that will be on it and
also absent the sports selector, which I actually forgot about.
After that, I will need to do a little bit of rejiggering to create
the controller and get it to do its thing.