Refactor View; Introduce Presenter
Does two different things, which is unfortunate in a commit. First
is that it refactors the way that the view works, so that common
features are combined in a way that I understand. It also introduces
the beginning of the MVP pattern with a Presenter. The View and the
Presenter are not connected as yet; in particular, the refactorings
of the View need to continue to make it match the interface that the
Presenter expects. When that is done, the pattern will probably
work in the way that I would expect it to.