Add initial test structure
Add some initial tests. Many of these are flat-out wrong, whether
they test the wrong things, or the things that they are supposed
to be testing are not well-exposed or as pure as they could be,
and many of the tests simply fail(). But this provides a foundation
for future work. One thing that is reasonably good is the
MainViewSpec, which uses TestFX as the test framework.