Add tests for the Model
This change involved a certain amount of work into
refactoring the model, which is principally just reponsible
for running the web scraper, to put it into a testable form.
The result is that the classes in the model are now, by and
large, set up to use dependency injection so that the
connection to a foreign URL can be mocked out.