Add seperate source set for UI tests
Add a seperate source set with Gradle (uiTest) to contain
the UI tests.  The principle reason for this is that it is a
workaround having to set up running the UI tests in headless
mode on a CI server (if I wind up doing that, which I
probably will). I will also likely get the tests to be run
in headless mode, but this is a useful intermediate step. It
also teaches me at least something about gradle. I will
probably also split the build.gradle file up at some point.