Add tests for presenter
Add tests for the MainPresenter class, which is responsible for
providing the callbacks which connect the view (which is, in this
case, mocked out) to the model (likewise). As there is only one
public method on the presenter, and it is void, this mostly
involved testing side effects caused by the callbacks on either the
view or on the model.