Seperate Gradle build script
Pull two logically distinct pieces of the gradle build into
script plugins, which made a little bit more sense to me.
In particular, it pulls out the section on UI tests (the
creation of the UI test source set and the TestFX runner)
and the section for the stylechecker into seperate .gradle
files.