Revise structure for UI tests
Prevent the UI tests from running by default - or indeed,
running at all outside of the IDE. The reason for this is
that Monocle is not in fact distributed with JavaFX in any
form, and would always need to be specifically built to be
used in tests. I don't really want to do that on my dev box
or on Travis, so that's out. The other change that this
makes is to make a test that wasn't passing pass. So that's
cool.