Initial code style changes
There are a lot of files touched by this change, but almost
all of the changes are meaningless. That is, most of these
changes are things like 4 space indentation to 2 space
indentation, or changing the order of imports. The goal is
to get it to (mostly) pass the checkstyle configuration.
Some things that will never pass: NU*.java, because I am not
going to rename those classes.