Drastically reduce the size of test HTML files
Had been using close to literal reproductions of the HTML
files that were being served by the CAA site. These files
were enormous.  This was causing GitHub to incorrectly
report this repo as being an HTML project with a smidgen of
Java, rather than a Java project that just happened to
contain some HTML files. This removes almost all of the
bloat from the relevant HTML files, which should make
everything much nicer.