Bugfix: Correct results display for unplayed games
Correct the display for unplayed games. Previously if a game
was not played it would render the result as "L - ", which
is not terribly meaningful. This fixes it to show only the
empty string if the game has no yet been played, as there is
not meaningful information to display.