Make query on tab change
The way that it had worked, there were effectively two
sources of truth for the currentDisplayType parameter. The
variable was maintained by the view, but updated by the
presenter. This was not a terribly good pattern, and when
there were inconsistencies between the two bad things
happened, as was the case in this instance. In particular,
it was the case that changing tabs made a request on the tab
that had just been left.

The pattern, quite frankly, is still not terribly good. I
would like to move that variable over to the presenter and
keep it there, as the view doesn't actually care about it.
Currently the view just holds it and keeps it available for
the presenter. This seems to indicate that the variable
should just live in the presenter.