Add artifacts to be collected by CircleCI
Add (the actual) style reports and the generated HTML test
reports to the CircleCI artifacts. This may not work as well
as I would like; in particular, I don't know how well Circle
is going to handle relative directories and javascript and
so on in terms of linking this stuff together.