Correct the save location of HTML style files
I had been trying to concatenate all of the HTML style files
into a single output HTML file. It didn't work, first,
second it had moved the html file into a location that I
didn't expect it to be in, and third was ignoring two/thirds
of the output.  This wasn't good. This fix puts the HTML
checkstyle files back where they are supposed to be.