Move web request to seperate thread
Previously the request to the model (and then, often, to the internet)
was on the same thread as the UI. This made the UI feel very laggy - for
example, clicking to switch tabs would cause the application to pause
while the query ran, and then would only switch tabs after the request
had been responded to.  Now the request is off of the UI thread, so
things like changing tabs is not blocked by making an HTTP request. It
still loads only slowly but there may be nothing to be done about that.