Lengthen timeout on fragile test
Extend the sleep duration in some of the Presenter tests. It is taking
just a little bit longer than I would like it to for the Presenter
thread to communicate with the UI thread, when a thread gets spun off to
make a request to the model. This will make the test suite run ever so
slightly slower, which is hopefully will not be noticable. More
fundamentally, it doesn't resolve the fundamental problem, which I think
is that is that I have the wrong API here. So that's a problem.