Await initial load in test
I misunderstood which end of the test was causing the problem. The
problem in this case was not caused by the check happening too soon
after interacting with the UI, but rather with a different check being
made too soon after the UI was first loaded. This includes a delay
before the first check, which appears to resolve the problem.