Add AssertJ as assertions framework
Swap out Hamcrest for AssertJ. I kind-of prefer AssertJ, at least as a
first impression. Code completion feels (much) nicer, and the mapping
utilities (#extracting and #filtering) are pretty nice.  It also feels
(to me) that it would be pretty amenable to the way Kotlin handles
extension classes, if I were to ever gather the enthusiasm to do that
re-writing process.