Add Spek as alternative test option
This adds a single Spek test file, which only duplicates the tests found
in NUWebScraperTest, the pre-existing JUnit test file. I do really like
Spek, but I don't feel that it is worthwhile at the moment to swap all
of my test files over at the moment. It is possible I will at some
point, especially if I wind up re-writing more of this project in Kotlin
(which I may or may not do).