Add a pause before running some tests
There are some tests that are only conditionally failing; in particular,
these tests are passing on my local machine and on CircleCI, but failing
on Travis. It appears to be a threading issues, in particular that tests
that are being run against the fictitious UI are being run before the UI
is fully loaded. This may or may not fix that.