Add missing sports to application
The missing sports (that I care about) were only Field Hockey and
Softball, both of which are now added. The CAA also includes football
and lax, but Northeastern does not compete in either of the sports.
Northeastern does compete in Swimming, XC, and Track and Field, but the
CAA does not (because of the format) publish standings for those events,
which renders this application mostly moot.