Bugfix: Preserve selected sport on switching tabs
The previous, incorrect behavior was that each tab would track the item
in the combo box - the selected sport - seperately. But the items in the
table were being tracked together. That meant that it was entirely
possible for the standings for Field Hockey, say, to be displayed while
what was evidently the the selected sport was Baseball. This fixes that
behavior to keep the selected item and the table contents in-sync.