Correct declared group of application
I had misspelled the group of the application as
"com.com.github.PtrTeixeira". I suspect that this was a relic of a
problem where that was, also incorrectly, the name of the package of the
code. It doesn't really matter, as this isn't a library or otherwise
intended for distribution, but it a bit of embarassing silliness.