Add tests for new controller
Add some tests to make sure that the controller does what it was
supposed to be doing. It wasn't, as a matter of fact; there was a bug
related to tab-switching order, but it was made in two places (once in
the controller, once in the view) and cancelled each other out. It would
have - and did, in fact - make refactoring difficult, though. This also
includes a fix for that.