Use Kotlin data classes for data objects
I had two data classes that were being passed from the model to the
view, Model and Standings. Because the easy path for tables in JavaFX
requires a particular bean format, this meant that there was a certain
amount of implicit bleed from the front-end data formats to the back-end
data formats. A combination of TornadoFX and Kotlin data classes neatens
that up nicely.

Having said that, this change DOES NOT change the former usages in the
model itself. So now the model uses broken view code, but the view
doesn't. Fancy that.