Rewrite model to use new data classes
This change was roughly on the scope that I imagined it would be.  On
the whole, almost no classes in the model change because I renamed the
data classes to replace the now-deleted Java data class usages.  Of
course, that necessitated changes in the presenter and view classes,
which were using the old names for the Kotlin data classes.

The second half probably wasn't necessary, and involved deleting API's
that referenced the old data classes. These API's certainly could have
been kept, but there weren't being used. Out of a standard desire to
remove dead code I dropped them, but that really is only incidentally
related to the main action of this commit.