Remove unused code from view
There was a function in the tab classes which had been used to attach
the lists from the model to the tables in the view. This was
restructured to more directly bind the list from the models into the
table, so it was not necessary to use a seperate function to provide
this binding.