Finish refactoring
Apparently I never ran the test suite before pushing, but I could have
sworn that I had. It's entirely possible that I just dreamed that I did;
I was pretty tired.

In any case, there was an incomplete refactoring that had left a type
parameter on a function that no longer needed a type parameter, which
causing the code to fail to compile. This _should_ fix that, as it
simply removes the type parameter in question.