Version bumping
This is the biggest part of what I'm trying to do with the refresh.  It
is a much smaller rewrite than I've done previously - not because I
necessarily think that this is in a better state than it has been each
time I've rewritten this, but because I have other projects that I'd
rather do instead.

This particular commit bumps versions of everything that _isn't_
TornadoFX, because that will take a little more care.