Refactor view
Continuing to use TornadoFX, refactor the view to be based on
Observables rather than whatever the hell I was doing before.  I'm not a
huge fan of the observables-as-controls paradigm (eg, MobX, Cycle in the
js world), but I was raised on the Elm architecture, so what can ya say.
This is definitely cleaner than it was, so I'll take it.