Port standings tab to new format
Port the standings tab to be set up the same way as the schedule tab. As
a result, I could make the body view a little neater.  It still isn't
perfect - I want to remove some duplication and extract a view model -
but it is doing pretty cool at the moment.