Finish re-writing model in Kotlin
There is now no remaining Java code; everything (should) now be in
Kotlin. This was mostly small changes; there really wasn't a huge
distinction between a Java exception class and a Kotlin exception class.
It was also an opportunity to clean up some things that weren't really
being used, and to (try to) introduce coroutines.  Not really sure if
what I got it "better" in any sense, but that's life.