Migrate to all JUnit 5
JUnit 5 has everything that I want out of a test framework. Spek has its
own problems (though much less than the used to be, to be fair) and its
API isn't something that I want to bother with at the moment. So I'm
cutting back on all of the test jars that I don't need in vafor of just
using junit.