Add loading bar
Add a loading bar to the bottom corner of the application when the UI is
loading new data into the table. This makes the UI much more obviously
responsive, rather than just having it appear to not do anything and
suddenly load more things into the table. I would prefer it if the table
went away while loading; will look into it.